requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prettytable==3.10.2
prompt_toolkit==3.0.47
rich==13.7.1
ruff==0.5.6
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tifffile==2024.7.24
timm==1.0.8
tomlkit==0.12.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.5
traitlets==5.14.3
triton==2.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
wcwidth==0.2.13
websockets==12.0
zipp==3.19.2
